requirements.txt
anyio==4.4.0
atomicwrites==1.4.1
attrs==23.2.0
Babel==2.15.0
beancount==2.3.6
beancount-import==1.4.0
beautifulsoup4==4.12.3
blinker==1.8.2
bottle==0.12.25
cachetools==5.3.3
certifi==2024.7.4
chardet==5.2.0
charset-normalizer==3.3.2
cheroot==10.0.1
click==8.1.7
colorlog==6.8.2
fava==1.28
Flask==3.0.3
flask-babel==4.0.0
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.2
httplib2==0.22.0
idna==3.7
iniconfig==2.0.0
itsdangerous==2.2.0
jaraco.functools==4.0.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
lxml==5.2.2
markdown2==2.5.0
MarkupSafe==2.1.5
more-itertools==10.3.0
nltk==3.8.1
numpy==2.0.0
packaging==24.1
pandas==2.2.2
pdfminer2==20151206
pluggy==1.5.0
ply==3.11
proto-plus==1.24.0
protobuf==5.27.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyparsing==3.1.2
pytest==8.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-magic==0.4.27
pytz==2024.1
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rpds-py==0.19.0
rsa==4.9
scikit-learn==1.5.1
scipy==1.14.0
simplejson==3.19.2
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
threadpoolctl==3.5.0
tornado==6.4.1
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
watchdog==4.0.1
watchfiles==0.22.0
Werkzeug==3.0.3
